Move the HTCondor compatibility launcher under executors/htcondor
