+++ executors/__init__.py
@@ -1,0 +1,1 @@
+"""Scheduler integrations, one package per Snakemake executor plugin (see docs/guide.md)."""
