+++ executors/htcondor/__init__.py
@@ -1,0 +1,1 @@
+"""HTCondor executor integration: profiles, job wrapper, and the submit-side launcher."""
