+++ executors/htcondor/snakemake_htcondor_compat.py
